Move Identifier class to its own file
This class is used in several locations and should be easy to include
anywhere without including the whole AST definitions.

gcc/rust/ChangeLog:

	* ast/rust-ast.h (class Identifier): Move from here to
	rust-ast-identifiers.h.
	(operator<<): Likewise.
	* ast/rust-ast-identifier.h: New file.
